+++ packages/pigeon/test/swift_generator_test.dart
@@ -395,7 +395,7 @@ void main() {
     );
     final String code = sink.toString();
     expect(code,
-        contains('completion: @escaping (Result<Void, FlutterError>) -> Void'));
+        contains('completion: @escaping (Result<Void, PigeonError>) -> Void'));
     expect(code, contains('completion(.success(Void()))'));
     expect(code, isNot(contains('if (')));
   });
@@ -477,7 +477,7 @@ void main() {
     expect(
         code,
         contains(
-            'func doSomething(completion: @escaping (Result<Output, FlutterError>) -> Void)'));
+            'func doSomething(completion: @escaping (Result<Output, PigeonError>) -> Void)'));
     expect(code, contains('channel.sendMessage(nil'));
     expect(code, isNot(contains('if (')));
   });
@@ -964,7 +964,7 @@ void main() {
     expect(
         code,
         contains(
-            'func doit(completion: @escaping (Result<[Int64?], FlutterError>) -> Void)'));
+            'func doit(completion: @escaping (Result<[Int64?], PigeonError>) -> Void)'));
     expect(code, contains('let result = listResponse[0] as! [Int64?]'));
     expect(code, contains('completion(.success(result))'));
   });
@@ -1052,7 +1052,7 @@ void main() {
     expect(
         code,
         contains(
-            'func add(x xArg: Int64, y yArg: Int64, completion: @escaping (Result<Int64, FlutterError>) -> Void)'));
+            'func add(x xArg: Int64, y yArg: Int64, completion: @escaping (Result<Int64, PigeonError>) -> Void)'));
     expect(code,
         contains('channel.sendMessage([xArg, yArg] as [Any?]) { response in'));
   });
@@ -1192,7 +1192,7 @@ void main() {
     expect(
         code,
         contains(
-            'func doit(foo fooArg: Int64?, completion: @escaping (Result<Void, FlutterError>) -> Void)'));
+            'func doit(foo fooArg: Int64?, completion: @escaping (Result<Void, PigeonError>) -> Void)'));
   });
 
   test('nonnull fields', () {
@@ -1572,6 +1572,6 @@ void main() {
     expect(
         code,
         contains(
-            'return FlutterError(code: "channel-error", message: "Unable to establish connection on channel: \'\\(channelName)\'.", details: "")'));
+            'return PigeonError(code: "channel-error", message: "Unable to establish connection on channel: \'\\(channelName)\'.", details: "")'));
   });
 }