+++ packages/go_router_builder/CHANGELOG.md
@@ -1,6 +1,7 @@
-## NEXT
+## 2.7.0
 
-* Updates minimum supported SDK version to Flutter 3.16/Dart 3.2.
+- Adds an example and a test with `onExit`.
+- Updates minimum supported SDK version to Flutter 3.16/Dart 3.2.
 
 ## 2.6.2